add many logs

+++ source/cmd/main.go
@@ -51,7 +51,7 @@ type Daftar struct {
 func databaseClose(db *gorm.DB) {
 	sqlDB, err := db.DB()
 	if err != nil {
-		log.Printf("Error Database Close %s", err.Error())
+		log.Printf("ABS01-Error Database Close %s", err.Error())
 	}
 	sqlDB.Close()
 }
@@ -61,16 +61,16 @@ func insertIntoDB(filename string, daftar *Daftar) error {
 	defer databaseClose(db)
 
 	if err != nil {
-		log.Panic("failed to connect database")
+		log.Panic("ABS02-failed to connect database")
 		return err
 	}
 	result := db.Create(daftar)
 	if err := result.Error; err != nil {
-		log.Panic("record couldn't be saved")
+		log.Panic("ABS03-record couldn't be saved")
 		return err
 	}
 	if count := result.RowsAffected; count == 0 {
-		log.Panic("record couldn't be saved")
+		log.Panic("ABS04-record couldn't be saved")
 		return err
 	}
 	return nil
@@ -81,7 +81,7 @@ func doLoginOrLogoutDB(filename string, isLogin bool) error {
 	defer databaseClose(db)
 
 	if err != nil {
-		log.Panic("failed to connect database")
+		log.Panic("ABS05-failed to connect database")
 		return err
 	}
 	daftar := &Daftar{}
@@ -103,16 +103,15 @@ func checkAndCreateSchedule(filename string) (*Daftar, error) {
 	defer databaseClose(db)
 
 	if err != nil {
-		log.Panic("failed to connect database")
+		log.Panic("ABS06-failed to connect database")
 		return nil, err
 	}
 	daftar := &Daftar{}
 	//db.First(&daftar, "workingDate=?", time.Now().Format("2006-01-02")).Error
-	log.Printf("Saved for Data: %s", time.Now().Format("2006-01-02"))
+	log.Printf("ABS07-Saved for Data: %s", time.Now().Format("2006-01-02"))
 	result := db.Where("working_date=?", time.Now().Format("2006-01-02")).First(&daftar)
 	if err = result.Error; errors.Is(err, gorm.ErrRecordNotFound) {
-		// Create another record
-		//log.Panicf("Create new record. %s", err.Error())
+		log.Printf("This action will insert new Punchin / Punchout record to local database")
 		loginTime, logoutTime := getDateLoginLogout()
 		daftar = &Daftar{
 			WorkingDate:    time.Now().Format("2006-01-02"),
@@ -129,31 +128,19 @@ func createFile(filePath string) error {
 	defer databaseClose(db)
 
 	if err != nil {
-		log.Panic("failed to connect database")
+		log.Panic("ABS08-failed to connect database")
 		return err
 	}
 	db.AutoMigrate(&Daftar{})
 	return nil
 }
-func doCron(delay int, config *config.Config) {
-	log.Printf("do Absensi at %s", time.Now())
+func doCron(delay int, config *ConfigFile) {
+	log.Printf("Last looping at %s and this is normal. Make sure this still looping.", time.Now())
 	s, err := scheduler.NewScheduler(1000)
 	if err != nil {
-		log.Println("Error when doing DoCron. Message: " + err.Error())
+		log.Println("ABS09-Error when doing DoCron. Message: " + err.Error())
 	}
 	s.Delay().Second(delay).Do(doAbsensi, config)
-}
-
-func copyToAbsensiDirectory(config *config.Config) error {
-	if !tools.CheckFileExists(".env") {
-		err := errors.New("cant find .env fle")
-		return err
-	}
-	if !tools.CheckFileExists(config.Picture) {
-		err := errors.New("cant find fle file")
-		return err
-	}
-	return nil
 }
 
 type ConfigFile struct {
@@ -177,15 +164,15 @@ func validateInput(value, validationTag string) string {
 		fmt.Printf("%s : ", value)
 		inputResult, err := reader.ReadString('\n')
 		if err != nil {
-			fmt.Printf("Error Gather Input %s\n", err)
+			fmt.Printf("ABS10-Error Gather Input %s\n", err)
 		}
 		inputResult = strings.Trim(inputResult, "\n")
 		// fmt.Printf("Input = %s\n", inputResult)
 		err = validate.Var(inputResult, validationTag)
 		if err != nil {
-			fmt.Printf("%s hanya bisa %s. %s\n", value, validationTag, err)
+			fmt.Printf("ABS11-%s hanya bisa %s. %s\n", value, validationTag, err)
 		} else if value == "Picture" && !tools.CheckFileExists(inputResult) {
-			fmt.Printf("File %s tidak ada\n", inputResult)
+			fmt.Printf("ABS12-File %s tidak ada\n", inputResult)
 		} else {
 			return inputResult
 		}
@@ -204,93 +191,55 @@ func gatherUserInput() *ConfigFile {
 	configFile.Lattitude = validateInput("Lattitude", "required,latitude")
 	configFile.Description = validateInput("Description", "required")
 	configFile.Region = "Asia/Jakarta"
-	configFile.BaseURL = "https://myapps.lintasarta.net/api"
+	configFile.BaseURL = "https://myapps.lintasarta.net/api/"
 	return configFile
 }
 
-func ValidateStruct(v ConfigFile) (err error) {
-	validate = validator.New()
-	err = validate.Struct(&v)
-	if err != nil {
-		if _, ok := err.(*validator.InvalidValidationError); ok {
-			fmt.Println(err)
-			return
-		}
-		// invalidErr := err.(*validator.InvalidValidationError)
-		// if invalidErr != nil {
-		// 	return
-		// }
-		// if _, ok := err.(*validator.InvalidValidationError); ok {
-		// 	fmt.Println(err)
-		// 	return
-		// }
-		// for _, err := range err.(validator.ValidationErrors) {
-
-		// 	fmt.Println(err.Namespace())
-		// 	fmt.Println(err.Field())
-		// 	fmt.Println(err.StructNamespace())
-		// 	fmt.Println(err.StructField())
-		// 	fmt.Println(err.Tag())
-		// 	fmt.Println(err.ActualTag())
-		// 	fmt.Println(err.Kind())
-		// 	fmt.Println(err.Type())
-		// 	fmt.Println(err.Value())
-		// 	fmt.Println(err.Param())
-		// 	fmt.Println()
-		// }
-		// return
-	}
-	return nil
-}
-
-// func validateConfig(sl validator.StructLevel) {
-
-// 	configFile := sl.Current().Interface().(ConfigFile)
-
-// 	// plus can do more, even with different tag than "fnameorlname"
-// }
-
 func main() {
+	configFile := NewConfigFile()
 	if !tools.CheckFileExists("config.json") {
-		configFile := gatherUserInput()
+		configFile = gatherUserInput()
 		file, _ := json.MarshalIndent(configFile, "", " ")
 		_ = ioutil.WriteFile("config.json", file, 0644)
 	} else {
-		configFile := NewConfigFile()
 		jsonFile, err := os.Open("config.json")
 		if err != nil {
-			log.Fatalf("File config.json tidak ditemukan. %s", err)
+			log.Fatalf("ABS13-File config.json tidak ditemukan. %s", err)
 		}
-		fmt.Println("Successfully Opened config.json")
+		// fmt.Println("Successfully Opened config.json")
 		defer jsonFile.Close()
 		byteValue, err := ioutil.ReadAll(jsonFile)
 		if err != nil {
-			log.Fatalf("File config.json corrupt. %s", err)
+			log.Fatalf("ABS14-File config.json corrupt. %s", err)
 		}
 		json.Unmarshal(byteValue, &configFile)
 		validate = validator.New()
 		err = validate.Struct(configFile)
 		if err != nil {
-			fmt.Println("File config.json missing information. Please check config.json. Error(s) are ")
+			fmt.Println("ABS15-File config.json missing information. Please check config.json. Error(s) are ")
 			errs := err.(validator.ValidationErrors)
 			for _, value := range errs {
 				if value.Tag() == "latitude" || value.Tag() == "longitude" {
-					fmt.Printf("Longitude or Latitude required or incorrect format\n")
+					fmt.Printf("ABS16-Longitude or Latitude required or incorrect format\n")
 				} else {
-					fmt.Printf("%s is %s\n", value.Field(), value.Tag())
+					fmt.Printf("ABS17-%s is %s\n", value.Field(), value.Tag())
 				}
 			}
 			return
 		}
 	}
-	config := config.NewConfig()
-	doAbsensi(config)
+	token := Login(configFile)
+	if token == "" {
+		fmt.Printf("ABS23-Your login is invalid. The apps will close")
+		return
+	}
+	doAbsensi(configFile)
 	sig := make(chan os.Signal)
 	signal.Notify(sig, os.Interrupt, os.Kill)
 	<-sig
 }
 
-func doAbsensi(config *config.Config) {
+func doAbsensi(config *ConfigFile) {
 	format := "2006-01-02 15:04"
 	home, _ := tools.GetHomeDirectory()
 	folderPath := home + "/" + ".absensi"
@@ -299,14 +248,14 @@ func doAbsensi(config *config.Config) {
 	if err != nil {
 		err = os.Mkdir(folderPath, os.ModePerm)
 		if err != nil {
-			log.Panic("Create File Error. Panic and abort the apps")
+			log.Panic("ABS18-Create File Error. Panic and abort the apps")
 		}
 	}
 	_, err = os.Stat(fileName)
 	if err != nil {
 		err = createFile(fileName)
 		if err != nil {
-			log.Panic("Create File Error. Panic and abort the apps")
+			log.Panic("ABS19-Create File Error. Panic and abort the apps")
 		}
 	}
 	currentDay := time.Now().Format("Mon")
@@ -321,9 +270,9 @@ func doAbsensi(config *config.Config) {
 		log.Debugf("bukan termasuk hari libur %t", existHariLibur)
 		daftar, err := checkAndCreateSchedule(fileName)
 		if err != nil {
-			log.Panic("Create File Error. Panic and abort the apps")
+			log.Panic("ABS20-Create File Error. Panic and abort the apps")
 		}
-		log.Printf("Login: %s,Logout: %s,Now: %s",
+		log.Printf("(Repeat) Login: %s,Logout: %s,Now: %s",
 			daftar.LoginDateTime.Format(format),
 			daftar.LogoutDateTime.Format(format),
 			time.Now().Format(format))
@@ -341,7 +290,7 @@ func doAbsensi(config *config.Config) {
 	doCron(59, config)
 }
 
-func doLoginOrLogout(config *config.Config, token string, isLogin bool) {
+func doLoginOrLogout(config *ConfigFile, token string, isLogin bool) {
 	url := ""
 	if isLogin {
 		url = config.BaseURL + "presences/punchIn"
@@ -356,11 +305,6 @@ func doLoginOrLogout(config *config.Config, token string, isLogin bool) {
 	_ = writer.WriteField("location_longitude", config.Longitude)
 	_ = writer.WriteField("location_description", config.Description)
 	_ = writer.WriteField("location_timezone", config.Region)
-	// home, err := tools.GetHomeDirectory()
-	// if err != nil {
-	// 	log.Panic("Create File Error. Panic and abort the apps")
-	// }
-	// folderPath := home + "/" + ".absensi"
 	file, errFile5 := os.Open(config.Picture)
 	if errFile5 != nil {
 		fmt.Println(errFile5)
@@ -418,7 +362,7 @@ func getDateLoginLogout() (time.Time, time.Time) {
 	randomLogoutShift := returnRandom(startRandomLogoutShift, 2)
 	fullLoginTime := loginTime.Add(time.Minute * time.Duration(randomLoginMinutes+(randomLoginShift*30)))
 	fullLogoutTime := logoutTime.Add(time.Minute * time.Duration(randomLogoutMinutes+(randomLogoutShift*30)))
-	log.Printf("Login: %s,Logout: %s", fullLoginTime, fullLogoutTime)
+	log.Printf("(Normal) Login: %s,Logout: %s", fullLoginTime, fullLogoutTime)
 	return fullLoginTime, fullLogoutTime
 }
 
@@ -426,17 +370,17 @@ func returnRandom(min, max int) int {
 	rand.Seed(time.Now().UnixNano())
 	x := rand.Intn(max-min+1) + min
 	// fmt.Println("Random %i", x)
-	log.Debugf("Random %d", x)
+	log.Debugf("ABS21-Random %d", x)
 	return x
 }
 
-func Login(config *config.Config) string {
+func Login(config *ConfigFile) string {
 	var params = url.Values{}
-	params.Add("username", config.Username)
+	params.Add("username", config.UserName)
 	params.Add("password", config.Password)
 	requestResult := tools.PostForm(config.BaseURL+"login", params)
 	if requestResult == "" {
-		log.Fatal("Error on Request Response")
+		log.Fatal("ABS22-Error on Request Response")
 	}
 	tokenResult := &TokenResult{}
 	tools.FromJSON(requestResult, tokenResult)
